feat(frontend): rename single dashboard list page
